refactor: 메뉴 갯수 렌더해주는 기능 render 폴더로 옮김
