+++ src/js/index.js
@@ -1,4 +1,4 @@
-import { renderMenuCount } from "./renderMenuCount.js"
+import { renderMenuCount } from "./render/renderMenuCount.js"
 import { renderMenuList } from "./render/renderMenuList.js";
 import { resetValue } from './util/resestValue.js'
 import { $ } from './util/selector.js'
